Remove verification from pallet
