+++ bin/runtime/src/lib.rs
@@ -342,7 +342,6 @@ impl pallet_aleph::Config for Runtime {
 parameter_types! {
     // We allow 10kB keys, proofs and public inputs. This is a 100% blind guess.
     pub const MaximumVerificationKeyLength: u32 = 10_000;
-    pub const MaximumDataLength: u32 = 10_000;
     pub const VerificationKeyDepositPerByte: u128 = MILLI_AZERO;
 }
 
@@ -352,7 +351,6 @@ impl pallet_baby_liminal::Config for Runtime {
     type RuntimeEvent = RuntimeEvent;
     type WeightInfo = pallet_baby_liminal::AlephWeight<Runtime>;
     type MaximumVerificationKeyLength = MaximumVerificationKeyLength;
-    type MaximumDataLength = MaximumDataLength;
     type VerificationKeyDepositPerByte = VerificationKeyDepositPerByte;
 }
