+++ pallets/baby-liminal/Cargo.toml
@@ -10,33 +10,20 @@ log = { version = "0.4", default-features = false }
 codec = { package = "parity-scale-codec", version = "3.0", default-features = false, features = ["derive"] }
 scale-info = { version = "2.0", default-features = false, features = ["derive"] }
 
-ark-bls12-381 = { version = "0.4.0", default-features = false, features = ["curve"] }
-ark-serialize = { version = "0.4.0", default-features = false }
-
 frame-benchmarking = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0", optional = true }
 frame-support = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 frame-system = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 sp-std = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 
-# we are waiting for: https://github.com/EspressoSystems/jellyfish/issues/258 and https://github.com/EspressoSystems/jellyfish/issues/268
-jf-plonk = { git = "https://github.com/Cardinal-Cryptography/jellyfish", branch = "substrate-compatible", default-features = false }
-
 [dev-dependencies]
-once_cell = { version = "1.17.1" }
 pallet-balances = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 sp-core = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 sp-io = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 sp-runtime = { default-features = false, git = "https://github.com/Cardinal-Cryptography/polkadot-sdk.git", branch = "aleph-v1.2.0" }
 
-jf-plonk = { git = "https://github.com/Cardinal-Cryptography/jellyfish", branch = "substrate-compatible", default-features = false, features = ["test-srs"] }
-jf-relation = { git = "https://github.com/Cardinal-Cryptography/jellyfish", branch = "substrate-compatible", default-features = false }
-jf-utils = { git = "https://github.com/Cardinal-Cryptography/jellyfish", branch = "substrate-compatible", default-features = false }
-
 [features]
 default = ["std"]
 std = [
-    "ark-bls12-381/std",
-    "ark-serialize/std",
     "log/std",
 
     "codec/std",
@@ -49,8 +36,6 @@ std = [
     "sp-std/std",
     "sp-runtime/std",
     "pallet-balances/std",
-
-    "jf-plonk/std",
 ]
 runtime-benchmarks = [
     "frame-benchmarking/runtime-benchmarks",
