+++ pallets/baby-liminal/src/tests/mod.rs
@@ -1,4 +1,3 @@
-mod relation;
 mod setup;
 mod suite;
